services/image_embedding/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
sentence-transformers
pillow
